fixes identity color calculation
previously, RGB values 0-15 were mapped to index 0 in the LUT.
Therefore, the darkest entries in the LUT were the average of 0-15, on
the order of 8 (usually more, due to the tone map), leaving shadows
too bright. A similar process limited highlight brightness.

This commit fixes the indexing, which now maps 0-7 to 0, 8-23 to 1,
229-245 to 14, and 245-255 to 15. This is probably a much more correct
way of doing things, and looks better, too.

A number of error corrections were removed, as they are no longer needed.